fix/refactor: remove Connections, rely on DependencyGraph
By relying on `Connections` for reading inputs, we would get an error when
data was produced via `outputs`.

This refactor moves all the deserialization logic into `Config`, using the
temporary structs `UserConfig` and `UserNodeConfig` to deserialize, and
keeping the `Deserialize` trait away from the `crate::nodes::*`
implementations.

This reduces the amount of boilerplate, makes the traversal of inputs more
consistent at nodes, and exposes less data to the nodes themselves, making
them more self-contained and functional.

Also, the main driver of the filter no longer relies on the node's cooperation
to get the node's name, fetching it directly via the configuration, ensuring
consistency and further reducing boilerplate.