fix(call): support degraded behavior in late responses
If the `response` node needs the `service_response_body` input,
it can only be triggered at the `HttpResponseBody` phase (which
in Nginx means `body_filter`). This means that the headers have
already been sent, meaning we can't use `ctx.send_http_response`.

By detecting this late phase, we degrade the behavior of the
node and only set the response body. More specifically, in
those cases the `response` node cannot set headers or a custom status.