fix: send a default error response if any node fails
Rather than ignoring failures, stop executing further nodes.
(In the future, we will have error catching and handling, but
for now it is better to stop the pipeline than to continue
proxying silently with an incomplete execution.)

The default behavior is to trigger an error response.
If debug tracing is enabled, the response error is not generated,
so that the execution trace can be printed in the response body.