fix(filter): always set service_request data
The host may not trigger `on_http_request_body` when the request
does not have a body (e.g. on `GET`), and the relevant data for
service_request_headers or `service_request_body` may be generated
late (i.e. after a `call`), so attempt to set those values on
at `on_http_call_response` as well.